Slight AINB nodes rework, fix some warnings

+++ src/ainb_editor/ainb_editor.hpp
@@ -21,7 +21,7 @@ private:
     ed::EditorContext *edContext = nullptr;
     std::vector<AINBImGuiNode> guiNodes;
     ed::NodeId rightClickedNode = 0;
-    int selectedNodeIdx = -1;
+    size_t selectedNodeIdx = -1;
     std::string selectedCommand = "";
 
     void AutoLayout();
